+++ src/zandronum/src/botpath.cpp
@@ -448,12 +448,12 @@ ULONG BOTPATH_TryWalk( AActor *pActor, fixed_t StartX, fixed_t StartY, fixed_t S
 	if ( AbsXDistance > AbsYDistance )
 	{
 		if ( AbsXDistance > MAXMOVE )
-			lNumSteps = 1 + ( AbsXDistance / MAXMOVE );
+			lNumSteps = (LONG)(1 + ( AbsXDistance / MAXMOVE ));
 	}
 	else
 	{
 		if ( AbsYDistance > MAXMOVE )
-			lNumSteps = 1 + ( AbsYDistance / MAXMOVE );
+			lNumSteps = (LONG)(1 + ( AbsYDistance / MAXMOVE ));
 	}
 
 	lCurrentStep = 1;
@@ -543,7 +543,7 @@ ULONG BOTPATH_TryWalk( AActor *pActor, fixed_t StartX, fixed_t StartY, fixed_t S
 			}
 		}
 
-		lHeightChange = g_PathSectorFloorZ - pActor->z;
+		lHeightChange = (LONG)(g_PathSectorFloorZ - pActor->z);
 		if ( lHeightChange > 0 )
 		{
 			if ( lHeightChange <= 0 /*gameinfo.StepHeight*/ )
@@ -622,7 +622,7 @@ ULONG BOTPATH_TryWalk( AActor *pActor, fixed_t StartX, fixed_t StartY, fixed_t S
 					pBackSector = pLine->frontsector;
 				}
 
-				lHeightChange = pBackSector->floorplane.ZatPoint( pLine->v1 ) - pFrontSector->floorplane.ZatPoint( pLine->v1 );
+				lHeightChange = (LONG)(pBackSector->floorplane.ZatPoint( pLine->v1 ) - pFrontSector->floorplane.ZatPoint( pLine->v1 ));
 				if ( lHeightChange > 0 )
 				{
 					if ( lHeightChange <= 0 /*gameinfo.StepHeight*/ )
