+++ src/zandronum/src/g_hexen/a_dragon.cpp
@@ -69,7 +69,7 @@ static void DragonSeek (AActor *actor, angle_t thresh, angle_t turnMax)
 		target->z+target->height < actor->z)
 	{
 		dist = (int)(P_AproxDistance(target->x-actor->x, target->y-actor->y));
-		dist = dist/actor->Speed;
+		dist = (int)(dist/actor->Speed);
 		if (dist < 1)
 		{
 			dist = 1;
@@ -79,7 +79,7 @@ static void DragonSeek (AActor *actor, angle_t thresh, angle_t turnMax)
 	else
 	{
 		dist = (int)(P_AproxDistance (target->x-actor->x, target->y-actor->y));
-		dist = dist/actor->Speed;
+		dist = (int)(dist/actor->Speed);
 	}
 	// [BB] If we're the server, update the thing's velocity and angle.
 	// Unfortunately there are sync issues, if we don't also update the actual position.
